feat: show ProcessMonitor loading errors
Normally missing qml modules like `QtWebSockets` are caught during widget initialization, but I am dynamically loading components depending on whether the optional C++ plugin is installed or not, and this way just throws a warning when the widget is added to a panel or when plasma starts.

Now these errors are shown on the widget.

refs: https://github.com/luisbocanegra/kurve/issues/29